move grpc to resource package
