+++ pkg/apiserver/go.mod
@@ -15,6 +15,7 @@ require (
 	k8s.io/apiserver v0.29.2
 	k8s.io/client-go v0.29.3
 	k8s.io/component-base v0.29.2
+	k8s.io/klog v1.0.0
 	k8s.io/klog/v2 v2.120.1
 	k8s.io/kube-openapi v0.0.0-20240228011516-70dd3763d340
 )
